+++ include/hep/mc/global_configuration.hpp
@@ -3,7 +3,7 @@
 
 /*
  * hep-mc - A Template Library for Monte Carlo Integration
- * Copyright (C) 2014  Christopher Schwan
+ * Copyright (C) 2014-2015  Christopher Schwan
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -24,33 +24,6 @@ namespace hep
 
 /// \addtogroup global_configuration
 /// @{
-
-/**
- * This function returns `true` the MPI functions use a single random number generator. This means
- * that the generator of each processor discards an appropriate amount of random numbers to
- * ascertain independent numbers are used. This make sure the result is independent of the number
- * of processesor and the same as the non-MPI routine. Since each process consumes as many random
- * numbers as calls are made in total, this may slow down the integration if the integrand evaluates
- * too fast. In this case a faster random number generator should be used.
- */
-inline bool& mpi_single_generator()
-{
-	static bool use_single_generator = false;
-
-	return use_single_generator;
-}
-
-/**
- * If `enabled` is `true` the MPI integration routines use a single random number generator so that
- * results are independent from the number of processors, in particular the same result than the
- * non-MPI integrators yield.
- *
- * \see \ref mpi_single_generator
- */
-inline void mpi_single_generator(bool enabled)
-{
-	mpi_single_generator() = enabled;
-}
 
 /**
  * Returns `true` if the VEGAS routines should use the grid refinement method of CUBA \cite Cuba for