+++ include/hep/mc/mpi_vegas.hpp
@@ -3,7 +3,7 @@
 
 /*
  * hep-mc - A Template Library for Monte Carlo Integration
- * Copyright (C) 2013-2014  Christopher Schwan
+ * Copyright (C) 2013-2015  Christopher Schwan
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -58,14 +58,6 @@ inline std::vector<vegas_iteration_result<T>> mpi_vegas(
 	int world = 0;
 	MPI_Comm_size(communicator, &world);
 
-	if (!mpi_single_generator())
-	{
-		// seed every random number generator differently
-		std::size_t const r = rank * 10;
-		std::seed_seq sequence{r+0, r+1, r+2, r+3, r+4, r+5, r+6, r+7, r+8, r+9};
-		generator.seed(sequence);
-	}
-
 	// create a fresh grid
 	auto pdf = start_pdf;
 
@@ -73,22 +65,18 @@ inline std::vector<vegas_iteration_result<T>> mpi_vegas(
 	std::vector<vegas_iteration_result<T>> results;
 	results.reserve(iteration_calls.size());
 
+	std::size_t const usage = pdf.dimensions() * random_number_usage<T, R>();
+
 	// perform iterations
 	for (auto i = iteration_calls.begin(); i != iteration_calls.end(); ++i)
 	{
-		if (mpi_single_generator())
-		{
-			advance_generator_before<T>(pdf.dimensions(), *i, rank, world, generator);
-		}
+		generator.discard(usage * discard_before(*i, rank, world));
 
 		std::size_t const calls = (*i / world) +
 			(static_cast <std::size_t> (rank) < (*i % world) ? 1 : 0);
 		auto result = vegas_iteration(calls, *i, pdf, function, generator);
 
-		if (mpi_single_generator())
-		{
-			advance_generator_after<T>(pdf.dimensions(), *i, calls, rank, world, generator);
-		}
+		generator.discard(usage * discard_after(*i, calls, rank, world));
 
 		// add up results
 		MPI_Allreduce(